fix(events-targets): add LogGroupTargetInput.fromObjectV2() method (#33720)
### Reason for this change

The `LogGroupTargetInput.fromObject()` function returns a `RuleTargetInput` instead of a `LogGroupTargetInput`. It can therefore not be passed into the following field:

```ts
export interface LogGroupProps extends TargetBaseProps {
  // ...
  readonly logEvent?: LogGroupTargetInput;
}
```
Because the types don't match.

This is a bug that hasn't shown up in TypeScript before because according to TypeScript's *structural typing* rules, the types `RuleTargetInput` and `LogGroupTargetInput` are compatible, but according to Java's *nominal typing* rules they are not.

### Description of changes

Deprecated `LogGroupTargetInput.fromObject()` and created `LogGroupTargetInput.fromObjectV2()` which returns the correct type of `LogGroupTargetInput`.

### Description of how you validated changes

integ test

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*