fix(core): remove whitespaces in tree.json (#33784)
### Issue # (if applicable)

Relates to #27261

### Reason for this change

Makes it possible to synthesize our CDK application, which errors due to tree.json hitting 512mb, which is node's max string size.

### Description of changes

tree.json can grow too large to be serialized, as detailed in #27261. One reason why the file is very large is that the serialization uses indentation space of 2. Removing the whitespace from serialization will return a lot of spaces, and it is a very straightforward change. This is OK to do because the file is not meant to be read by human anyway. This PR implements the solution.

Ideally, in the future we should also make it possible to split tree.json into multiple files or use the solution in https://github.com/bmoffatt/aws-cdk/commit/91edfec0365ab5ff2f479df675d9afc6f3a54fab, to make it scalable. At that stage, it's also possible to turn indentation space back to 2 (perhaps to aid debugging).

For now, however, this issue is blocking the deployment in our team, and we really need this fix to continue.

### Describe any new or updated permissions being added

No permission involved.


### Description of how you validated changes

Does not affect any functionality. 

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*